Much more stable
GPT 3 is extremely bad at following instructions and not getting distracted. I now made the script make it clear to GPT 3 how it's supposed to behave, by example. I also had to make a content filtering mechanism so that if GPT 3 outputs something that isn't a terminal command (which it often does because it's bad at following instructions) then I don't put that into the history, because then the bad output from GPT distracts GPT in future requests.